+++ yazi-binding/src/runtime.rs
@@ -1,5 +1,6 @@
 use std::{collections::VecDeque, mem};
 
+use anyhow::{Context, Result};
 use hashbrown::{HashMap, hash_map::EntryRef};
 use mlua::Function;
 
@@ -11,12 +12,12 @@ pub struct Runtime {
 }
 
 #[derive(Debug)]
-struct RuntimeFrame {
+pub struct RuntimeFrame {
 	id: String,
 }
 
 impl Runtime {
-	pub fn new() -> Self { Self { frames: <_>::default(), blocks: <_>::default(), blocking: true } }
+	pub fn new() -> Self { Self { frames: <_>::default(), blocks: <_>::default(), blocking: false } }
 
 	pub fn new_isolate(id: &str) -> Self {
 		Self {
@@ -28,28 +29,32 @@ impl Runtime {
 
 	pub fn push(&mut self, id: &str) { self.frames.push_back(RuntimeFrame { id: id.to_owned() }); }
 
-	pub fn pop(&mut self) { self.frames.pop_back(); }
+	pub fn pop(&mut self) -> Result<RuntimeFrame> {
+		self.frames.pop_back().context("Runtime stack underflow")
+	}
 
 	pub fn critical_push(&mut self, id: &str, blocking: bool) -> bool {
 		self.push(id);
 		mem::replace(&mut self.blocking, blocking)
 	}
 
-	pub fn critical_pop(&mut self, blocking: bool) {
-		self.pop();
+	pub fn critical_pop(&mut self, blocking: bool) -> Result<RuntimeFrame> {
 		self.blocking = blocking;
+		self.pop()
 	}
 
-	pub fn current(&self) -> Option<&str> { self.frames.back().map(|f| f.id.as_str()) }
+	pub fn current(&self) -> Result<&str> {
+		self.frames.back().map(|f| f.id.as_str()).context("No current runtime frame")
+	}
 
-	pub fn current_owned(&self) -> Option<String> { self.current().map(ToOwned::to_owned) }
+	pub fn current_owned(&self) -> Result<String> { self.current().map(ToOwned::to_owned) }
 
 	pub fn get_block(&self, id: &str, calls: usize) -> Option<Function> {
 		self.blocks.get(id).and_then(|v| v.get(calls)).cloned()
 	}
 
 	pub fn put_block(&mut self, f: &Function) -> Option<usize> {
-		let Some(cur) = self.frames.back() else { return None };
+		let Some(cur) = self.frames.back().filter(|f| f.id != "init") else { return None };
 		Some(match self.blocks.entry_ref(&cur.id) {
 			EntryRef::Occupied(mut oe) => {
 				oe.get_mut().push(f.clone());