+++ yazi-plugin/src/utils/sync.rs
@@ -1,5 +1,3 @@
-use std::mem;
-
 use anyhow::Context;
 use futures::future::join_all;
 use mlua::{ExternalError, ExternalResult, Function, IntoLuaMulti, Lua, MultiValue, Table, Value, Variadic};
@@ -16,27 +14,26 @@ use crate::loader::LOADER;
 impl Utils {
 	pub(super) fn sync(lua: &Lua) -> mlua::Result<Function> {
 		lua.create_function(|lua, f: Function| {
-			let Some(block) = runtime_mut!(lua)?.put_block(&f) else {
+			let mut rt = runtime_mut!(lua)?;
+			let Some(block) = rt.put_block(&f) else {
 				return Err("`ya.sync()` must be called in a plugin").into_lua_err();
 			};
 
+			let current = rt.current_owned()?;
 			lua.create_async_function(move |lua, mut args: MultiValue| {
-				let f = f.clone();
+				let (f, current) = (f.clone(), current.clone());
 				async move {
-					let rt = runtime!(lua)?;
-					let (Some(cur), blocking) = (rt.current_owned(), rt.blocking) else {
-						return Err("`ya.sync()` block must be used within a plugin").into_lua_err();
-					};
-					drop(rt);
-
+					let blocking = runtime!(lua)?.blocking;
 					if blocking {
-						args.push_front(Value::Table(LOADER.try_load(&lua, &cur)?));
+						args.push_front(Value::Table(LOADER.try_load(&lua, &current)?));
 						f.call::<MultiValue>(args)
 					} else {
-						Self::retrieve(&lua, &cur, block, args)
+						Self::retrieve(&lua, &current, block, args)
 							.await
 							.and_then(|data| Sendable::list_to_values(&lua, data))
-							.with_context(|| format!("Failed to execute sync block-{block} in `{cur}` plugin"))
+							.with_context(|| {
+								format!("Failed to execute sync block-{block} in `{current}` plugin")
+							})
 							.into_lua_err()
 					}
 				}
@@ -51,26 +48,18 @@ impl Utils {
 			})
 		} else {
 			lua.create_function(|lua, (f, args): (Function, MultiValue)| {
-				let name = runtime!(lua)?.current_owned();
+				let name = runtime!(lua)?.current_owned()?;
 				let lua = lua.clone();
 
 				Ok(Handle::AsyncFn(LOCAL_SET.spawn_local(async move {
-					let blocking = mem::replace(&mut runtime_mut!(lua)?.blocking, false);
-					if let Some(s) = &name {
-						runtime_mut!(lua)?.push(s); // make `current_owned` always return Some to eliminate the check
-					}
-
+					let blocking = runtime_mut!(lua)?.critical_push(&name, false);
 					let result = f.call_async::<MultiValue>(args).await;
-
-					runtime_mut!(lua)?.blocking = blocking;
-					if name.is_some() {
-						runtime_mut!(lua)?.pop();
-					}
+					runtime_mut!(lua)?.critical_pop(blocking)?;
 
 					if let Err(ref e) = result {
-						match name {
-							Some(s) => tracing::error!("Failed to execute async block in `{s}` plugin: {e}"),
-							None => tracing::error!("Failed to execute async block in `init.lua`: {e}"),
+						match name.as_str() {
+							"init" => tracing::error!("Failed to execute async block in `init.lua`: {e}"),
+							s => tracing::error!("Failed to execute async block in `{s}` plugin: {e}"),
 						}
 					}
 					result