+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/):
 
 ### Added
 
+- Allow using `ps.sub()` in `init.lua` directly without a plugin ([#3638])
 - New `cx.which` API to access the which component state ([#3617])
 - New `ind-which-activate` DDS event to change the which component behavior ([#3608])
 
@@ -1634,3 +1635,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/):
 [#3617]: https://github.com/sxyazi/yazi/pull/3617
 [#3633]: https://github.com/sxyazi/yazi/pull/3633
 [#3634]: https://github.com/sxyazi/yazi/pull/3634
+[#3638]: https://github.com/sxyazi/yazi/pull/3638